# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SerWish API
  version: 2.1.0
  description: |
    API for the SerWish customer and partner apps.

    **Envelope.** Every JSON response is `{ success, message, data?, meta? }` on success and
    `{ success: false, code, message, errors?, data? }` on failure. `errors` lists `{ field, message }`
    for `VALIDATION_FAILED`; `data` carries context such as the fresh estimate for `PRICE_CHANGED`.

    **Auth.** Sign in with Firebase in the app (Google first; phone OTP; email and password for
    Play Store testers), then send the Firebase ID token as `Authorization: Bearer <token>`.
    Call `POST /auth/session` once after each sign-in.

    **Status.** `x-status: live` endpoints work today. `phase-5` (online payments) and `phase-6`
    (KYC upload, earnings) endpoints answer `501 NOT_IMPLEMENTED` until that backend phase ships.

    **Money** is whole rupees (integers). **Times** are ISO 8601 UTC.
servers:
  - url: http://localhost:5000/api/v1
    description: Local
  - url: https://api.serwish.in/api/v1
    description: Production (planned)
security:
  - firebase: []
...
paths:
...
  /providers/{id}/reviews:
    get:
      tags: [Partners]
      x-status: live
      summary: Reviews for a partner, newest first
      parameters: [{ $ref: '#/components/parameters/PartnerId' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Offset' }]
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/Ok' }, { properties: { data: { type: array, items: { $ref: '#/components/schemas/Review' } }, meta: { $ref: '#/components/schemas/PageMeta' } } }] } } }
...
  /bookings/{id}/review:
    post:
      tags: [Bookings]
      x-status: live
      summary: Rate a completed booking (optional, once). Alias POST /reviews { bookingId, rating, comment }.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                text: { type: string, maxLength: 500 }
                tags: { type: array, maxItems: 6, items: { type: string } }
      responses: { '201': { description: Saved }, '409': { description: '`ALREADY_REVIEWED` or not completed yet' } }
...
  /notifications:
    get:
      tags: [Account]
      x-status: live
      summary: My notifications, newest first (meta.unread = unread count)
      parameters:
        - { name: unread, in: query, schema: { type: boolean } }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/Ok' }, { properties: { data: { type: array, items: { $ref: '#/components/schemas/Notification' } } } }] } } } }
  /notifications/read:
    post:
      tags: [Account]
      x-status: live
      summary: Mark all (or the given ids) as read
      requestBody:
        content: { application/json: { schema: { type: object, properties: { ids: { type: array, items: { type: string, format: uuid } } } } } }
      responses: { '200': { description: OK } }
...
components:
  securitySchemes:
    firebase:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  parameters:
...
    PartnerId: { name: id, in: path, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_-]{1,128}$' }, description: Partner user id (Firebase UID) }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
    Offset: { name: offset, in: query, schema: { type: integer, minimum: 0, maximum: 10000, default: 0 } }
...
  schemas:
    Ok:
      type: object
      required: [success, message]
      properties:
        success: { const: true }
        message: { type: string }
...
    PageMeta:
      type: object
      properties: { total: { type: integer }, limit: { type: integer }, offset: { type: integer } }
...
    Review:
      type: object
      properties:
        id: { type: string, format: uuid }
        author: { type: string }
        avatar: { type: [string, 'null'] }
        date: { type: string, format: date-time }
        rating: { type: integer, minimum: 1, maximum: 5 }
        text: { type: string }
        tags: { type: array, items: { type: string } }
        photos: { type: array, items: { type: string } }
        service: { type: [string, 'null'] }
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [booking, offer, update, job] }
        title: { type: string }
        body: { type: string }
        at: { type: string, format: date-time }
        read: { type: boolean }
        bookingId: { type: string }
        icon: { type: string, enum: [check, onTheWay, offer, completed, feature, gift, cancelled] }
...